move the Modal editor out of the block instance
